+++ src/css/mod.rs
@@ -1,9 +1,11 @@
 //! Types for css values and rules.
 mod call_args;
 mod rule;
+mod string;
 mod value;
 mod valueformat;
 
 pub use self::call_args::CallArgs;
 pub use self::rule::{BodyItem, Rule};
+pub use self::string::CssString;
 pub use self::value::{Value, ValueMap};