+++ src/css/valueformat.rs
@@ -1,29 +1,13 @@
 use super::Value;
 use crate::output::Formatted;
-use crate::value::{ListSeparator, Operator, Quotes};
+use crate::value::{ListSeparator, Operator};
 use std::fmt::{self, Display, Write};
 
 impl<'a> Display for Formatted<'a, Value> {
     fn fmt(&self, out: &mut fmt::Formatter) -> fmt::Result {
         match *self.value {
             Value::Bang(ref s) => write!(out, "!{}", s),
-            Value::Literal(ref s, ref q) => match *q {
-                Quotes::None => write!(out, "{}", s),
-                Quotes::Double => {
-                    if s.contains('"') && !s.contains('\'') {
-                        write_sq(out, s)
-                    } else {
-                        write_dq(out, s)
-                    }
-                }
-                Quotes::Single => {
-                    if !s.contains('"') || s.contains('\'') {
-                        write_dq(out, s)
-                    } else {
-                        write_sq(out, s)
-                    }
-                }
-            },
+            Value::Literal(ref s) => s.fmt(out),
             Value::Function(ref n, ref _f) => {
                 let name = n
                     .chars()
@@ -172,25 +156,4 @@ impl<'a> Display for Formatted<'a, Value> {
             }
         }
     }
-}
-
-fn write_dq(out: &mut fmt::Formatter, s: &str) -> fmt::Result {
-    out.write_char('"')?;
-    for c in s.chars() {
-        if c == '"' {
-            out.write_char('\\')?;
-        }
-        out.write_char(c)?;
-    }
-    out.write_char('"')
-}
-fn write_sq(out: &mut fmt::Formatter, s: &str) -> fmt::Result {
-    out.write_char('\'')?;
-    for c in s.chars() {
-        if c == '\'' {
-            out.write_char('\\')?;
-        }
-        out.write_char(c)?;
-    }
-    out.write_char('\'')
 }